+++ appseed/aura/primitive/geometry/geometry.inl
@@ -66,4 +66,4 @@ inline point null_point() { return point(0, 0); }
 
 
 inline pointd::pointd(const SIZE & sz) NOTHROW : pointd(sz.cx, sz.cy) {}
-inline pointd::pointd(const SIZE64 & sz) NOTHROW : pointd(sz.cx, sz.cy) {}
+inline pointd::pointd(const SIZE64 & sz) NOTHROW : pointd((double)sz.cx, (double)sz.cy) {}
